+++ feature/library/src/main/kotlin/com/kelsos/mbrc/feature/library/genres/BaseGenreViewModel.kt
@@ -42,4 +42,8 @@ abstract class BaseGenreViewModel(
       emit(message)
     }
   }
+
+  fun goToAlbums(genre: Genre) {
+    launchDefault(GenreUiMessage.OpenAlbums(genre))
+  }
 }