+++ feature/library/src/main/kotlin/com/kelsos/mbrc/feature/library/genres/GenreUiMessage.kt
@@ -6,6 +6,8 @@ import com.kelsos.mbrc.core.data.library.genre.Genre
 sealed class GenreUiMessage : UiMessageBase {
   data class OpenArtists(val genre: Genre) : GenreUiMessage()
 
+  data class OpenAlbums(val genre: Genre) : GenreUiMessage()
+
   data class QueueSuccess(val tracksCount: Int) : GenreUiMessage()
 
   object QueueFailed : GenreUiMessage()